feat: update feature komik cmd

+++ src/commands/komik/komik-delete.ts
@@ -14,26 +14,25 @@ const response = {
     const messageUser = deletePrefixCommand(message.message?.conversation?.toLowerCase() as string, this.name);
 
     const userDataDB = await prisma.user.findMany({ where: { number: keyRemoteJid } });
-    if (!userDataDB.length)
-      return await sock.sendMessage(keyRemoteJid, { text: `Kamu belum terdaftar.\nSilahkan daftar dengan *${config.prefix}komik-regis*` });
+    if (!userDataDB.length) {
+      return sock.sendMessage(keyRemoteJid, { text: `Kamu belum terdaftar.\nSilahkan daftar dengan *${config.prefix}komik-regis*` });
+    }
+
+    const komikDataDB = await prisma.subscription_komik.findMany({ where: { user_id: userDataDB[0].id } });
+    if (!komikDataDB.length) {
+      return sock.sendMessage(keyRemoteJid, {
+        text: `Kamu belum berlangganan komik.\nSilahkan berlangganan *${config.prefix}komik-subs [nama komik]*`,
+      });
+    }
 
     try {
       const deleteKomikDB = await prisma.subscription_komik.delete({
-        where: {
-          id: userDataDB[0].id,
-          komik_name: {
-            contains: messageUser,
-          },
-        },
+        where: { user_id: userDataDB[0].id, komik_name: { contains: messageUser } },
       });
-
-      if (!deleteKomikDB)
-        return await sock.sendMessage(keyRemoteJid, { text: `Komik *${messageUser}* tidak ditemukan.\nSilahkan list *${config.prefix}komik-list*` });
-
-      return await sock.sendMessage(keyRemoteJid, { text: `Berhasil menghapus komik *${messageUser}* dari daftar` });
+      return sock.sendMessage(keyRemoteJid, { text: `Berhasil menghapus komik *${deleteKomikDB.komik_name}* dari daftar` });
     } catch (error) {
       console.log(error);
-      return await sock.sendMessage(keyRemoteJid, { text: `Terjadi kesalahan saat menghapus data komik.` });
+      return sock.sendMessage(keyRemoteJid, { text: `Terjadi kesalahan saat menghapus data komik.` });
     }
   },
 };
